backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
click==8.1.7
dlib==19.24.2
dnspython==2.6.1
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
Flask==3.0.2
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
motor==3.3.2
pyasn1==0.5.1
pydantic==2.6.2
pydantic_core==2.16.3
pymongo==4.6.2
python-dateutil==2.8.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing==3.7.4.3
typing_extensions==4.10.0
uvicorn==0.27.1
Werkzeug==3.0.1
zipp==3.17.0
